requirements.txt
llama-index
google-generativeai
llama-index-llms-google-genai
pypdf
python-dotenv
Ipython
llama-index-embeddings-google-genai
streamlit